Small correction
Compose Desktop - for now - is also built using the JVM, then wrapped in a "native distribution" (i.e. the final packages to install). That's why we use Java libraries within the desktop sources.

There are some efforts to also make it use Kotlin/Native (with some experimental work already done for macOS), but it's still far from ready, and not a high priority thing for now.

This commit corrects that part.